allow dynamic block data in TerraScript block function by default
