--- common/addons/structure-terrascript-loader/src/main/java/com/dfsek/terra/addons/terrascript/script/functions/AbstractBlockFunction.java
@@ -1,54 +1,0 @@
-package com.dfsek.terra.addons.terrascript.script.functions;
-
-import com.dfsek.terra.addons.terrascript.buffer.items.BufferedBlock;
-import com.dfsek.terra.addons.terrascript.parser.lang.ImplementationArguments;
-import com.dfsek.terra.addons.terrascript.parser.lang.Returnable;
-import com.dfsek.terra.addons.terrascript.parser.lang.functions.Function;
-import com.dfsek.terra.addons.terrascript.parser.lang.variables.Variable;
-import com.dfsek.terra.addons.terrascript.script.TerraImplementationArguments;
-import com.dfsek.terra.addons.terrascript.tokenizer.Position;
-import com.dfsek.terra.api.TerraPlugin;
-import com.dfsek.terra.api.block.state.BlockState;
-import com.dfsek.terra.api.util.RotationUtil;
-import com.dfsek.terra.api.vector.Vector2;
-import com.dfsek.terra.api.vector.Vector3;
-import net.jafama.FastMath;
-
-import java.util.Map;
-
-public abstract class AbstractBlockFunction implements Function<Void> {
-    protected final Returnable<Number> x, y, z;
-    protected final Returnable<String> blockData;
-    protected final TerraPlugin main;
-    private final Returnable<Boolean> overwrite;
-    private final Position position;
-
-    protected AbstractBlockFunction(Returnable<Number> x, Returnable<Number> y, Returnable<Number> z, Returnable<String> blockData, Returnable<Boolean> overwrite, TerraPlugin main, Position position) {
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.blockData = blockData;
-        this.overwrite = overwrite;
-        this.main = main;
-        this.position = position;
-    }
-
-    void setBlock(ImplementationArguments implementationArguments, Map<String, Variable<?>> variableMap, TerraImplementationArguments arguments, BlockState rot) {
-        Vector2 xz = new Vector2(x.apply(implementationArguments, variableMap).doubleValue(), z.apply(implementationArguments, variableMap).doubleValue());
-
-        RotationUtil.rotateVector(xz, arguments.getRotation());
-
-        RotationUtil.rotateBlockData(rot, arguments.getRotation().inverse());
-        arguments.getBuffer().addItem(new BufferedBlock(rot, overwrite.apply(implementationArguments, variableMap), main, arguments.isWaterlog()), new Vector3(FastMath.roundToInt(xz.getX()), y.apply(implementationArguments, variableMap).doubleValue(), FastMath.roundToInt(xz.getZ())));
-    }
-
-    @Override
-    public Position getPosition() {
-        return position;
-    }
-
-    @Override
-    public ReturnType returnType() {
-        return ReturnType.VOID;
-    }
-}
