+++ common/addons/structure-terrascript-loader/src/main/java/com/dfsek/terra/addons/terrascript/script/functions/BlockFunction.java
@@ -1,36 +1,66 @@
 package com.dfsek.terra.addons.terrascript.script.functions;
 
-import com.dfsek.terra.addons.terrascript.parser.exceptions.ParseException;
+import com.dfsek.terra.addons.terrascript.buffer.items.BufferedBlock;
 import com.dfsek.terra.addons.terrascript.parser.lang.ImplementationArguments;
 import com.dfsek.terra.addons.terrascript.parser.lang.Returnable;
-import com.dfsek.terra.addons.terrascript.parser.lang.constants.ConstantExpression;
+import com.dfsek.terra.addons.terrascript.parser.lang.functions.Function;
 import com.dfsek.terra.addons.terrascript.parser.lang.variables.Variable;
 import com.dfsek.terra.addons.terrascript.script.TerraImplementationArguments;
 import com.dfsek.terra.addons.terrascript.tokenizer.Position;
 import com.dfsek.terra.api.TerraPlugin;
 import com.dfsek.terra.api.block.state.BlockState;
+import com.dfsek.terra.api.util.RotationUtil;
+import com.dfsek.terra.api.vector.Vector2;
+import com.dfsek.terra.api.vector.Vector3;
+import net.jafama.FastMath;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 
-public class BlockFunction extends AbstractBlockFunction {
-    private final BlockState data;
+public class BlockFunction implements Function<Void> {
+    protected final Returnable<Number> x, y, z;
 
-    public BlockFunction(Returnable<Number> x, Returnable<Number> y, Returnable<Number> z, Returnable<String> data, Returnable<Boolean> overwrite, TerraPlugin main, Position position) throws ParseException {
-        super(x, y, z, data, overwrite, main, position);
+    private final Map<String, BlockState> data = new HashMap<>();
+    protected final Returnable<String> blockData;
+    protected final TerraPlugin main;
+    private final Returnable<Boolean> overwrite;
+    private final Position position;
 
-        if(!(data instanceof ConstantExpression)) throw new ParseException("Block data must be constant", data.getPosition());
-        try {
-            this.data = main.getWorldHandle().createBlockData(((ConstantExpression<String>) data).getConstant());
-        } catch(IllegalArgumentException e) {
-            throw new ParseException("Could not parse block data", data.getPosition(), e);
-        }
+    public BlockFunction(Returnable<Number> x, Returnable<Number> y, Returnable<Number> z, Returnable<String> blockData, Returnable<Boolean> overwrite, TerraPlugin main, Position position) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+        this.blockData = blockData;
+        this.overwrite = overwrite;
+        this.main = main;
+        this.position = position;
+    }
+
+    void setBlock(ImplementationArguments implementationArguments, Map<String, Variable<?>> variableMap, TerraImplementationArguments arguments, BlockState rot) {
+        Vector2 xz = new Vector2(x.apply(implementationArguments, variableMap).doubleValue(), z.apply(implementationArguments, variableMap).doubleValue());
+
+        RotationUtil.rotateVector(xz, arguments.getRotation());
+
+        RotationUtil.rotateBlockData(rot, arguments.getRotation().inverse());
+        arguments.getBuffer().addItem(new BufferedBlock(rot, overwrite.apply(implementationArguments, variableMap), main, arguments.isWaterlog()), new Vector3(FastMath.roundToInt(xz.getX()), y.apply(implementationArguments, variableMap).doubleValue(), FastMath.roundToInt(xz.getZ())));
     }
 
     @Override
     public Void apply(ImplementationArguments implementationArguments, Map<String, Variable<?>> variableMap) {
         TerraImplementationArguments arguments = (TerraImplementationArguments) implementationArguments;
-        BlockState rot = data.clone();
+        BlockState rot = data.computeIfAbsent(blockData.apply(implementationArguments, variableMap), main.getWorldHandle()::createBlockData).clone();
         setBlock(implementationArguments, variableMap, arguments, rot);
         return null;
+    }
+
+    @Override
+    public Position getPosition() {
+        return position;
+    }
+
+    @Override
+    public ReturnType returnType() {
+        return ReturnType.VOID;
     }
 }
